Changed handling of device data in views
